// admin/src/lib/types.ts
export type RequestQuery<T> = {
    page?: number | null
    sort?: string | null
    limit?: number | null
    fields?: string | null
} & Partial<T>

export interface IBaseModel {
    _id: string
    createdAt: Date
    updatedAt: Date
}
// ...
export interface ISiteAnalytics extends IBaseModel {
    date: Date
    totalVisits: number
    newSignups: number
    geoData: { country: string; visits: number }[]
    deviceData: { deviceType: string }[]
}

export interface IOrderAnalytics extends IBaseModel {
    date: Date
    totalOrders: number
    revenue: number
    averageOrderValue: number
}